Refactor: enums.py
