+++ src/unimport/commands/parser.py
@@ -1,7 +1,8 @@
 import argparse
 
-from unimport import __description__, emoji
+from unimport import __description__
 from unimport.commands import options
+from unimport.enums import Emoji
 
 __all__ = ("generate_parser",)
 
@@ -10,7 +11,7 @@ def generate_parser() -> argparse.ArgumentParser:
     parser = argparse.ArgumentParser(
         prog="unimport",
         description=__description__,
-        epilog=f"Get rid of all unused imports {emoji.PARTYING_FACE}",
+        epilog=f"Get rid of all unused imports {Emoji.PARTYING_FACE}",
     )
     exclusive_group = parser.add_mutually_exclusive_group(required=False)
