+++ src/unimport/commands/permission.py
@@ -1,7 +1,8 @@
 from pathlib import Path
 
 from unimport import utils
-from unimport.color import Color, paint
+from unimport.color import paint
+from unimport.enums import Color
 
 __all__ = ("permission",)
