--- src/unimport/emoji.py
@@ -1,4 +1,0 @@
-__all__ = ("PARTYING_FACE", "STAR")
-
-STAR = "\U0001F929"
-PARTYING_FACE = "\U0001F973"
