--- tests/test_emoji.py
@@ -1,6 +1,0 @@
-from unimport import emoji
-
-
-def test_emoji():
-    assert emoji.STAR == "🤩"
-    assert emoji.PARTYING_FACE == "🥳"
