+++ tests/test_enum.py
@@ -1,0 +1,6 @@
+from unimport.enums import Emoji
+
+
+def test_emoji():
+    assert Emoji.STAR == "🤩"
+    assert Emoji.PARTYING_FACE == "🥳"
